hey remember to check that it actually broke

+++ src/main/java/io/github/afamiliarquiet/familiar_magic/block/SummoningTableBlock.java
@@ -184,13 +184,15 @@ public class SummoningTableBlock extends BaseEntityBlock {
 
     @Override
     protected void onRemove(BlockState state, Level level, BlockPos pos, BlockState newState, boolean movedByPiston) {
-        // tell the stuff in the inventory to bounce
-        BlockEntity blockEntity = level.getBlockEntity(pos);
-        if (blockEntity instanceof SummoningTableBlockEntity summoningTableInventory) {
-            for (int i = 0; i < summoningTableInventory.getSlots(); i++) {
-                Containers.dropItemStack(level, pos.getX(), pos.getY(), pos.getZ(), summoningTableInventory.getStackInSlot(i));
+        if (!state.is(newState.getBlock())) {
+            // tell the stuff in the inventory to bounce
+            BlockEntity blockEntity = level.getBlockEntity(pos);
+            if (blockEntity instanceof SummoningTableBlockEntity summoningTableInventory) {
+                for (int i = 0; i < summoningTableInventory.getSlots(); i++) {
+                    Containers.dropItemStack(level, pos.getX(), pos.getY(), pos.getZ(), summoningTableInventory.getStackInSlot(i));
+                }
             }
+            super.onRemove(state, level, pos, newState, movedByPiston);
         }
-        super.onRemove(state, level, pos, newState, movedByPiston);
     }
 }
